backend/requirement.txt
djangorestframework==3.15.1
django-cors-headers==4.3.1
django==5.0.4
pillow